backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
motor
pymongo
python-dotenv
email-validator
dnspython
